Break out IAPPTable to more generalized RecordTable
